% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_cluster.R
\name{fe_cluster}
\alias{fe_cluster}
\title{Feature Engineering Clustering}
\usage{
fe_cluster(DT, test, predInClust = 5)
}
\arguments{
\item{a}{train data table}

\item{a}{test data table}

\item{number}{of k in kmeans clustering}
}
\value{
a list with finished two data tables
}
\description{
This function takes in a data table and a desired number of features in a cluster. This function will filter only
numerical features first and cluster them with randomly picked numerical features. the final data table will have added
clustered id's and distance to centroids added
}
