% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_pca.R
\name{fe_pca}
\alias{fe_pca}
\title{Feature Engineering PCA vector addition}
\usage{
fe_pca(train, test, pca_var, pca_comp, mca_var, mca_comp)
}
\arguments{
\item{a}{train data table}

\item{a}{test data table}

\item{ratio}{of variance explained (pca)}

\item{a}{number of components to be used (pca)}

\item{ratio}{of variace explained (mca)}

\item{a}{number of components to be used (mca)}
}
\value{
a list with finished two data tables
}
\description{
This function will take in data and add pca feature engineering (PC components) to the original table. For categorical variables,
MCA will be applied for feature reduction.
}
