% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findUniqueKey.R
\name{findUniqueKey}
\alias{findUniqueKey}
\title{Finding unique key}
\usage{
findUniqueKey(DT)
}
\arguments{
\item{a}{data table}
}
\value{
a logical vector
}
\description{
This function takes a data table and returns a logical vectors telling which column has a unique identifier
}
