% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_encoding.R
\name{freq_encoding}
\alias{freq_encoding}
\title{frequency encoding}
\usage{
freq_encoding(data, column)
}
\arguments{
\item{a}{data table and a string of a column}
}
\value{
a data frame with frequency encoded values
}
\description{
This function takes a data table and target column and gives frequency encoding for the target column
}
