% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cat_data.R
\name{get_cat_data}
\alias{get_cat_data}
\title{Getting categorical data}
\usage{
get_cat_data(DT)
}
\arguments{
\item{a}{data table}
}
\value{
a data table with categorical features
}
\description{
This function takes a datatable and slices only categorical data
}
