% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_num_data.R
\name{get_num_data}
\alias{get_num_data}
\title{Getting numeric data}
\usage{
get_num_data(DT)
}
\arguments{
\item{a}{data table}
}
\value{
sliced data table with only numerical features
}
\description{
this function takes a datatable and filters only numeric data
}
