% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distance.R
\name{getDistance}
\alias{getDistance}
\title{Getting distance between two data points}
\usage{
getDistance(p1, p2)
}
\arguments{
\item{a}{point}

\item{a}{different point}
}
\value{
distance value
}
\description{
This function calculates euclidean distance between two points
}
