% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeVal.R
\name{imputeVal}
\alias{imputeVal}
\title{Impute Values}
\usage{
imputeVal(dt, method = "median")
}
\arguments{
\item{a}{Data Table with missing values}
}
\value{
A finished data table with filled missing data table
}
\description{
Objective is to find NA values and imput using median. for categorical values, "None" factor will be added in place for 
missing values.
}
