% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipak.R
\name{ipak}
\alias{ipak}
\title{loading necessary libraries}
\usage{
ipak(pkg)
}
\arguments{
\item{a}{character vector of library names}
}
\value{
boolean values of whether libraries loaded successfully
}
\description{
This function takes in a vector of libraries and load them to the environment
}
