% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_var_plot.R
\name{pca_varExp_plot}
\alias{pca_varExp_plot}
\title{PCA variance explanined plot}
\usage{
pca_varExp_plot(md.pca)
}
\arguments{
\item{a}{prcomp object}
}
\value{
a plot of variance explained
}
\description{
This function takes a model pca object and plots a pca variance explained plot
}
