% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_dt.R
\name{preprocess_dt}
\alias{preprocess_dt}
\title{Preprocessing data}
\usage{
preprocess_dt(df)
}
\arguments{
\item{a}{data table}
}
\value{
a data table
}
\description{
This function will eliminate zero variance column and unique identifier column
This is a preliminary step for fe_cluster
}
