% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_column.R
\name{slice_column}
\alias{slice_column}
\title{Slicing a data table with column names}
\usage{
slice_column(DT, columnNames)
}
\arguments{
\item{a}{data table}

\item{a}{string or a string vector with column name(s)}
}
\value{
a data table
}
\description{
This function does column slicing from a data table
}
