% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_mean_ec.R
\name{target_mean_enc}
\alias{target_mean_enc}
\title{Target mean encoding}
\usage{
target_mean_enc(data, column, target)
}
\arguments{
\item{a}{data table, column string, target string}
}
\value{
a data frame with resulted values
}
\description{
this function takes a data table and a column and a target predictor to calculate target mean encoding.
}
