% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroVar.R
\name{zeroVar}
\alias{zeroVar}
\title{Finding zero variance}
\usage{
zeroVar(DT)
}
\arguments{
\item{a}{data table}
}
\value{
a logical vector indicating which columns have one value for all
}
\description{
This function takes a data table and finds both numeric and character predictors that have only one value (zero variance predictors)
}
